+++ test/test_comm.py
@@ -7,7 +7,7 @@ from typing import Tuple
 import pytest
 
 from deisa.dask.communicator import DaskComm, resolve_comm
-from deisa.dask.types import DeisaArray
+import xarray as xr
 
 
 def mpi_gather_test():
@@ -145,8 +145,9 @@ def test_mpi_bridge(global_size: Tuple, parallelism: int, comm: str):
 
     deisa = Deisa(get_connection_info=lambda: client, wait_for_go=False)
     darr = deisa.get_array('temperature', iteration=1)
-    assert isinstance(darr, DeisaArray)
-    assert darr.dask.sum().compute() == np.prod(global_size), f"temperature sum should be the product of {global_size}"
+    assert isinstance(darr, xr.DataArray)
+    assert darr.coords["t"].item() == 1
+    assert darr.data.sum().compute() == np.prod(global_size), f"temperature sum should be the product of {global_size}"
 
     cluster.close()