+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Types of changes:
 - set Client `heartbeat_interval` to `sys.maxsize`
 - Change `TestSimulation` to use `asyncio` to run `bridge.send`
 - Gracefully stop bridges. Deisa waits for all bridges to close
+- Replace DeisaArray wrapper by Xarray
 
 ### Fixed